# [8.3.0] Optimize `--remote_download_regex` patterns