--- a/MODULE.bazel
+++ b/MODULE.bazel
@@ -378,6 +378,7 @@ use_repo(
 # =========================================
 
 bazel_dep(name = "bazel_ci_rules", version = "1.0.0")
+bazel_dep(name = "rules_fuzzing", version = "0.5.2")
 
 rbe_preconfig = use_repo_rule("@bazel_ci_rules//:rbe_repo.bzl", "rbe_preconfig")