… src/test/java/com/google/devtools/build/lib/util/regex/RegexUtilFuzzTest.java
… third_party/BUILD