Java's regex engine optimizes `find` but not `matches` operations, which can result in very slow matching behavior for `--remote_download_regex` patterns when applied to builds with large tree artifacts. In a large iOS build with five relatively simple patterns, `RemoteOutputChecker.matchesPattern` accounted for 11% of CPU time.

We fix this by rewriting regexes into an equivalent `find` pattern, which benefits from the Boyer-Moore optimization pass applied by the engine (~3-5x speedup on the benchmark). Along the way, patterns of the form `.*<literal>` are turned into extremely fast suffix matches (~17x speedup on the benchmark).

This requires a change in behavior that is incompatible in theory but negligible in practice: regex options now use `Pattern.DOTALL` and `Pattern.UNIX_LINES` semantics, which means that `.` also matches line terminators and `\n` is the only line terminator. For all Bazel flags that accept regexes, line terminators aren't relevant.

The new matcher is also used in other places that match a user-provided pattern against an entire string.

The change includes a benchmark and a fuzz test to verify correctness of the transformation.

```
Benchmark                                                                                                         (haystack)                            (needle)   Mode  Cnt          Score          Error  Units
baseline   bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer                    .*/coverage\.dat  thrpt   12    4172476.649 ±    52130.107  ops/s
baseline   bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer             .*/my_action.outputs/.*  thrpt   12    4162228.888 ±    54490.078  ops/s
baseline   bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer            .*/testlogs/.*/test\.xml  thrpt   12    4106391.495 ±    43602.839  ops/s
baseline   bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer  .*/testlogs/.*/attempt_[0-9]+\.xml  thrpt   12    4120051.290 ±    29027.736  ops/s
baseline        bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log                    .*/coverage\.dat  thrpt   12    4312175.136 ±    35139.695  ops/s
baseline        bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log             .*/my_action.outputs/.*  thrpt   12    4173191.063 ±   185695.084  ops/s
baseline        bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log            .*/testlogs/.*/test\.xml  thrpt   12    2112244.239 ±    12178.108  ops/s
baseline        bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log  .*/testlogs/.*/attempt_[0-9]+\.xml  thrpt   12    2122376.835 ±    11950.259  ops/s
optimized  bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer                    .*/coverage\.dat  thrpt   12  712157494.127 ± 10435487.808  ops/s
optimized  bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer             .*/my_action.outputs/.*  thrpt   12   20193918.187 ±    79263.024  ops/s
optimized  bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer            .*/testlogs/.*/test\.xml  thrpt   12   22143940.479 ±    65747.617  ops/s
optimized  bazel-out/darwin_arm64-opt-exec-ST-fad1763555eb/bin/src/main/java/com/google/devtools/build/lib/bazel/BazelServer  .*/testlogs/.*/attempt_[0-9]+\.xml  thrpt   12   22167011.261 ±    66956.978  ops/s
optimized       bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log                    .*/coverage\.dat  thrpt   12  718395295.848 ±  4945067.723  ops/s
optimized       bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log             .*/my_action.outputs/.*  thrpt   12   18364343.102 ±    44631.925  ops/s
optimized       bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log            .*/testlogs/.*/test\.xml  thrpt   12    5900847.315 ±    47399.997  ops/s
optimized       bazel-out/darwin_arm64-fastbuild/testlogs/src/test/java/com/google/devtools/common/options/AllTests/test.log  .*/testlogs/.*/attempt_[0-9]+\.xml  thrpt   12    6095470.481 ±    13954.493  ops/s
```

Closes #26056.

PiperOrigin-RevId: 759582918
Change-Id: Ied70b6c4669d72a6090ff3ec9b9832d9ff88865f

Commit https://github.com/bazelbuild/bazel/commit/662c0f0ed15a7e27f9c9caf08fc60af0fd993ccb